Move reducers to separate file per action
